fix wrongly moved files
